+++ imod/mf6/wel.py
@@ -33,7 +33,7 @@ from imod.mf6.mf6_wel_adapter import Mf6Wel, concat_indices_to_cellid
 from imod.mf6.package import Package
 from imod.mf6.utilities.dataset import remove_inactive
 from imod.mf6.utilities.imod5_converter import well_from_imod5_cap_data
-from imod.mf6.validation_context import ValidationContext
+from imod.mf6.validation_settings import ValidationSettings
 from imod.mf6.write_context import WriteContext
 from imod.prepare import assign_wells
 from imod.prepare.cleanup import cleanup_wel, cleanup_wel_layered
@@ -455,7 +455,7 @@ class GridAgnosticWell(BoundaryCondition, IPointDataPackage, abc.ABC):
         Mf6Wel
             Object with wells as list based input.
         """
-        validation_context = ValidationContext(
+        validation_context = ValidationSettings(
             validate=validate, strict_well_validation=strict_well_validation
         )
         return self._to_mf6_pkg(active, top, bottom, k, validation_context)
@@ -466,7 +466,7 @@ class GridAgnosticWell(BoundaryCondition, IPointDataPackage, abc.ABC):
         top: GridDataArray,
         bottom: GridDataArray,
         k: GridDataArray,
-        validation_context: ValidationContext,
+        validation_context: ValidationSettings,
     ) -> Mf6Wel:
         if validation_context.validate:
             errors = self._validate(self._write_schemata)